Fixed burger styling
