+++ components/index.js
@@ -9,8 +9,10 @@ import Accordion from "./Accordion";
 import AccordionSection from "./AccordionSection";
 import LetterSelectBar from "./LetterSelectBar";
 import Resources from "./Resources";
+import BurgerMenu from "./BurgerMenu";
 
 export {
+    BurgerMenu,
     InfoBar,
     PopUp,
     NavBar,
